feat: std::sync::mpsc single-thread bench + CLI subcommands (0.2.0-dev4)
First IIAC bench: StdMpscRoundTrip keeps Sender + Receiver on one
thread, sends a u64 then recv()s it. CLI gains subcommands: timer,
channel, all (default).

3900x: round-trip min 20 ns, p50 26 ns, p99 31 ns at 100M calls —
close to std::time::Instant::now() cost since single-thread mpsc has
no contention or thread blocking. std::sync::mpsc is implemented on
top of crossbeam-channel since Rust ~1.67, so this effectively
measures crossbeam through the std API.

Also widened Bench::step return type from u32 to u64 — drops the
as u32 truncation in channel.rs and fits future benches that
naturally produce u64.

ochid: /.claude/lyuwnmuoqtsn

+++ src/benches/timer.rs
@@ -10,7 +10,7 @@ impl Bench for MinstantNow {
         "minstant::Instant::now()"
     }
 
-    fn step(&mut self) -> u32 {
+    fn step(&mut self) -> u64 {
         black_box(minstant::Instant::now());
         1
     }
@@ -23,7 +23,7 @@ impl Bench for StdInstantNow {
         "std::time::Instant::now()"
     }
 
-    fn step(&mut self) -> u32 {
+    fn step(&mut self) -> u64 {
         black_box(std::time::Instant::now());
         1
     }
